+++ src/libaktualizr/primary/aktualizr.cc
@@ -69,9 +69,9 @@ bool Aktualizr::UptaneCycle() {
   return true;
 }
 
-std::future<void> Aktualizr::RunForever() {
-  std::future<void> future = std::async(std::launch::async, [&]() {
-    SendDeviceData().get();
+std::future<void> Aktualizr::RunForever(const Json::Value &custom_hwinfo) {
+  std::future<void> future = std::async(std::launch::async, [this, custom_hwinfo]() {
+    SendDeviceData(custom_hwinfo).get();
 
     std::unique_lock<std::mutex> l(exit_cond_.m);
     while (true) {
@@ -136,8 +136,8 @@ std::future<void> Aktualizr::CampaignControl(const std::string &campaign_id, cam
   return api_queue_.enqueue(task);
 }
 
-std::future<void> Aktualizr::SendDeviceData() {
-  std::function<void()> task([this] { uptane_client_->sendDeviceData(); });
+std::future<void> Aktualizr::SendDeviceData(const Json::Value &custom_hwinfo) {
+  std::function<void()> task([this, custom_hwinfo] { uptane_client_->sendDeviceData(custom_hwinfo); });
   return api_queue_.enqueue(task);
 }